Update pp.c
We are running Hashtopolis and some tasks with Prince activated. We have noticed that wordlists encoded in UTF8 create problems with characters such as åäö (swedish). The linux executable works fine but the windows executable, pp64.exe, does not output the characters correctly, this in turn means that all the agents running Hashtopolis in Windows will not be able to crack passwords with åäö.

Forcing UTF-8 during compile seemed to solve the problem, not sure if there is a mor elegant solution.